+++ NEWS.md
@@ -6,8 +6,9 @@
 - button to export the plot in "Run Model" tab (#9) and in "Post Processing" tab
 
 ### Updates
-- change the default value of "Select a file" in the "Run Model" tab to the last uploaded file
-(not the first upload) (#6)
+- in the "Run Model" tab:
+  - change the default value of "Select a file" to the last uploaded file (not the first upload) (#6)
+  - add a load button to load a selected file (#11)
 - update the UI:
   - remove right sidebar, 
   - integrate export buttons into main panel,
@@ -18,3 +19,4 @@
 - transform columns to numeric columns when running a model (#8), all NA's are removed automatically
 - fix wrong default values for width and height when exporting a plot (#9, #10)
 - fix `tryCatch` when fitting a model, error message is now forwarded correctly to a pop up (#9)
+- fix import/upload of files with the same name (#11)